+++ recognition/__init__.py
@@ -10,4 +10,4 @@
 """
 
 __all__ = ["template_matcher", "types"]
-__version__ = "0.2.0"
+__version__ = "0.7.0"