+++ state/__init__.py
@@ -8,4 +8,4 @@
 """
 
 __all__ = ["game_state", "types"]
-__version__ = "0.2.0"
+__version__ = "0.7.0"